[Chart] Move datapoint updating logic to modules/Chart
Chart, the rendering component, does not have any data-related logic at
this point then. It’s just a renderer – as should be!
